[a]delete_by_metadata refuses empty filters + adjusted tests

+++ libs/astradb/langchain_astradb/vectorstores.py
@@ -830,58 +830,60 @@ class AstraDBVectorStore(VectorStore):
     def delete_by_metadata_filter(
         self,
         filter: dict[str, Any],  # noqa: A002
-    ) -> int | None:
+    ) -> int:
         """Delete all documents matching a certain metadata filtering condition.
 
         This operation does not use the vector embeddings in any way, it simply
         removes all documents whose metadata match the provided condition.
-        Use with caution: passing an empty filter dictionary results in
-        completely emptying the vector store.
 
         Args:
-            filter: Filter on the metadata to apply.
+            filter: Filter on the metadata to apply. The filter cannot be empty.
 
         Returns:
             An number expressing the amount of deleted documents.
-            This will be None if a `{}` metadata filter condition is passed,
-            implying emptying the store entirely.
         """
+        if not filter:
+            msg = (
+                "Method `delete_by_metadata_filter` does not accept an empty "
+                "filter. Use the `clear()` method if you really want to empty "
+                "the vector store."
+            )
+            raise ValueError(msg)
         self.astra_env.ensure_db_setup()
         metadata_parameter = self.filter_to_query(filter)
         del_result = self.astra_env.collection.delete_many(
             filter=metadata_parameter,
         )
-        if del_result.deleted_count is not None and del_result.deleted_count >= 0:
-            return del_result.deleted_count
-        return None
+        return del_result.deleted_count or 0
 
     async def adelete_by_metadata_filter(
         self,
         filter: dict[str, Any],  # noqa: A002
-    ) -> int | None:
+    ) -> int:
         """Delete all documents matching a certain metadata filtering condition.
 
         This operation does not use the vector embeddings in any way, it simply
         removes all documents whose metadata match the provided condition.
-        Use with caution: passing an empty filter dictionary results in
-        completely emptying the vector store.
 
         Args:
-            filter: Filter on the metadata to apply.
+            filter: Filter on the metadata to apply. The filter cannot be empty.
 
         Returns:
             An number expressing the amount of deleted documents.
-            This will be None if a `{}` metadata filter condition is passed,
-            implying emptying the store entirely.
         """
+        if not filter:
+            msg = (
+                "Method `delete_by_metadata_filter` does not accept an empty "
+                "filter. Use the `clear()` method if you really want to empty "
+                "the vector store."
+            )
+            raise ValueError(msg)
         await self.astra_env.aensure_db_setup()
         metadata_parameter = self.filter_to_query(filter)
         del_result = await self.astra_env.async_collection.delete_many(
             filter=metadata_parameter,
         )
-        if del_result.deleted_count is not None and del_result.deleted_count >= 0:
-            return del_result.deleted_count
-        return None
+        return del_result.deleted_count or 0
 
     def delete_collection(self) -> None:
         """Completely delete the collection from the database.
@@ -1230,7 +1232,7 @@ class AstraDBVectorStore(VectorStore):
     ) -> int:
         """Add/overwrite the metadata of existing documents.
 
-        For each document to update, the new metadata dictionary is added
+        For each document to update, the new metadata dictionary is appended
         to the existing metadata, overwriting individual keys that existed already.
 
         Args:
@@ -1281,7 +1283,7 @@ class AstraDBVectorStore(VectorStore):
     ) -> int:
         """Add/overwrite the metadata of existing documents.
 
-        For each document to update, the new metadata dictionary is added
+        For each document to update, the new metadata dictionary is appended
         to the existing metadata, overwriting individual keys that existed already.
 
         Args:
